+++ stac_fastapi/extensions/stac_fastapi/extensions/core/__init__.py
@@ -5,6 +5,7 @@ from .collection_search import CollectionSearchExtension
 from .context import ContextExtension
 from .fields import FieldsExtension
 from .filter import FilterExtension
+from .free_text import FreeTextAdvancedExtension, FreeTextExtension
 from .pagination import PaginationExtension, TokenPaginationExtension
 from .query import QueryExtension
 from .sort import SortExtension
@@ -15,6 +16,8 @@ __all__ = (
     "ContextExtension",
     "FieldsExtension",
     "FilterExtension",
+    "FreeTextExtension",
+    "FreeTextAdvancedExtension",
     "PaginationExtension",
     "QueryExtension",
     "SortExtension",